refactor: update test imports to use dist instead of src
- Changed all test imports from src/ to dist/ directories
- Tests now run against compiled code rather than source TypeScript
- Ensures tests validate the actual published artifacts
- Updated core package tests (tinqer)
- Updated adapter package tests (pg-promise, better-sqlite3)
- Minor README formatting improvements for table alignment

This change makes tests more representative of the actual package behavior that consumers will experience.
