+++ packages/tinqer/tests/groupby.test.ts
@@ -5,7 +5,7 @@
 
 import { describe, it } from "mocha";
 import { expect } from "chai";
-import { parseQuery, from } from "../src/index.js";
+import { parseQuery, from } from "../dist/index.js";
 import {
   asGroupByOperation,
   asSelectOperation,
