+++ packages/tinqer/tests/parse-query-integration.test.ts
@@ -4,7 +4,7 @@
 
 import { describe, it } from "mocha";
 import { expect } from "chai";
-import { parseQuery, from } from "../src/index.js";
+import { parseQuery, from } from "../dist/index.js";
 import {
   asFromOperation,
   asWhereOperation,
@@ -13,14 +13,14 @@ import {
   asTakeOperation,
   getOperation,
 } from "./test-utils/operation-helpers.js";
-import type { ParamRef } from "../src/query-tree/operations.js";
+import type { ParamRef } from "../dist/query-tree/operations.js";
 import type {
   ComparisonExpression,
   ObjectExpression,
   LogicalExpression,
   BooleanMethodExpression,
   ColumnExpression,
-} from "../src/expressions/expression.js";
+} from "../dist/expressions/expression.js";
 import { db } from "./test-schema.js";
 
 describe("Parse Query Integration Tests", () => {
