+++ packages/tinqer/tests/select.test.ts
@@ -4,10 +4,10 @@
 
 import { describe, it } from "mocha";
 import { expect } from "chai";
-import { parseQuery, from } from "../src/index.js";
+import { parseQuery, from } from "../dist/index.js";
 import { db } from "./test-schema.js";
 import { asSelectOperation, getOperation } from "./test-utils/operation-helpers.js";
-import type { ObjectExpression, ColumnExpression } from "../src/expressions/expression.js";
+import type { ObjectExpression, ColumnExpression } from "../dist/expressions/expression.js";
 
 describe("SELECT Operation", () => {
   describe("Simple Projections", () => {
