+++ packages/tinqer/tests/skip.test.ts
@@ -4,15 +4,15 @@
 
 import { describe, it } from "mocha";
 import { expect } from "chai";
-import { parseQuery, from } from "../src/index.js";
+import { parseQuery, from } from "../dist/index.js";
 import {
   asSkipOperation,
   asOrderByOperation,
   asTakeOperation,
   getOperation,
 } from "./test-utils/operation-helpers.js";
-import type { ParamRef } from "../src/query-tree/operations.js";
-import type { ArithmeticExpression } from "../src/expressions/expression.js";
+import type { ParamRef } from "../dist/query-tree/operations.js";
+import type { ArithmeticExpression } from "../dist/expressions/expression.js";
 import { db } from "./test-schema.js";
 
 describe("SKIP Operations", () => {
