+++ packages/tinqer/tests/take.test.ts
@@ -4,7 +4,7 @@
 
 import { describe, it } from "mocha";
 import { expect } from "chai";
-import { parseQuery, from } from "../src/index.js";
+import { parseQuery, from } from "../dist/index.js";
 import {
   asTakeOperation,
   asWhereOperation,
@@ -12,7 +12,7 @@ import {
   asSelectOperation,
   getOperation,
 } from "./test-utils/operation-helpers.js";
-import type { ParamRef } from "../src/query-tree/operations.js";
+import type { ParamRef } from "../dist/query-tree/operations.js";
 import { db } from "./test-schema.js";
 
 describe("TAKE Operation", () => {
