+++ packages/tinqer/tests/where.test.ts
@@ -4,7 +4,7 @@
 
 import { describe, it } from "mocha";
 import { expect } from "chai";
-import { parseQuery, from } from "../src/index.js";
+import { parseQuery, from } from "../dist/index.js";
 import { db } from "./test-schema.js";
 import { expr } from "./test-utils/expr-helpers.js";
 import { asWhereOperation, getOperation } from "./test-utils/operation-helpers.js";
@@ -14,8 +14,8 @@ import type {
   ColumnExpression,
   ConstantExpression,
   ParameterExpression,
-} from "../src/expressions/expression.js";
-import type { ParamRef } from "../src/query-tree/operations.js";
+} from "../dist/expressions/expression.js";
+import type { ParamRef } from "../dist/query-tree/operations.js";
 
 describe("WHERE Operation", () => {
   describe("Comparison Operators", () => {
